* release/undev_v3.13.0: (427 commits)
  fix sattelites error
  Remove modernizr
  use last released gem elasticsearch-git
  fix = ""
  fix html escaping
  Fix {-Headers in push email
  Fix Dangerous settings button on edit team page
  fix 500 when update merge request
  Improve results count texts
  refs fixes #45242
  Fix autocomplete search
  Mass improve search
  Add search in description for Group, Team, Project
  Add . to indexing
  Allow _ to search
  fix search russian text
  simple improve search logic
  fix navigation
  fix project search
  fix nav for empty search query
  ...

+++ lib/gitlab/satellite/satellite.rb
@@ -1,5 +1,9 @@
 module Gitlab
-  class SatelliteNotExistError < StandardError;  end
+  class SatelliteNotExistError < StandardError
+    def initialize(msg = "Satellite doesn't exist")
+      super
+    end
+  end
 
   module Satellite
     class Satellite
@@ -17,14 +21,9 @@ module Gitlab
         Gitlab::Satellite::Logger.error(message)
       end
 
-      def raise_no_satellite
-        raise SatelliteNotExistError.new("Satellite doesn't exist")
-      end
-
       def clear_and_update!
-        raise_no_satellite unless exists?
+        raise SatelliteNotExistError unless exists?
 
-        File.exists? path
         @repo = nil
         clear_working_dir!
         delete_heads!
@@ -33,7 +32,7 @@ module Gitlab
       end
 
       def create
-        output, status = popen("git clone #{project.repository.path_to_repo} #{path}",
+        output, status = popen(%W(git clone -- #{project.repository.path_to_repo} #{path}),
                                Gitlab.config.satellites.path)
 
         log("PID: #{project.id}: git clone #{project.repository.path_to_repo} #{path}")
@@ -55,7 +54,7 @@ module Gitlab
       # * Changes the current directory to the satellite's working dir
       # * Yields
       def lock
-        raise_no_satellite unless exists?
+        raise SatelliteNotExistError unless exists?
 
         File.open(lock_file, "w+") do |f|
           begin
@@ -77,7 +76,7 @@ module Gitlab
       end
 
       def repo
-        raise_no_satellite unless exists?
+        raise SatelliteNotExistError unless exists?
 
         @repo ||= Grit::Repo.new(path)
       end
@@ -98,19 +97,25 @@ module Gitlab
       # This ensures we have no name clashes or issues updating branches when
       # working with the satellite.
       def delete_heads!
-        heads = repo.heads.map(&:name)
+        begin
+          heads = repo.heads.map(&:name)
 
-        # update or create the parking branch
-        if heads.include? PARKING_BRANCH
-          repo.git.checkout({}, PARKING_BRANCH)
-        else
-          repo.git.checkout(default_options({b: true}), PARKING_BRANCH)
+          # update or create the parking branch
+          if heads.include? PARKING_BRANCH
+            repo.git.checkout({}, PARKING_BRANCH)
+          else
+            repo.git.checkout(default_options({b: true}), PARKING_BRANCH)
+          end
+
+          # remove the parking branch from the list of heads ...
+          heads.delete(PARKING_BRANCH)
+          # ... and delete all others
+          heads.each { |head| repo.git.branch(default_options({D: true}), head) }
+        rescue
+          destroy
+          create
+          clear_and_update!
         end
-
-        # remove the parking branch from the list of heads ...
-        heads.delete(PARKING_BRANCH)
-        # ... and delete all others
-        heads.each { |head| repo.git.branch(default_options({D: true}), head) }
       end
 
       # Deletes all remotes except origin
